add PUTSPRITE
