+++ examples/Example02/MAKEFILE.BAT
@@ -3,8 +3,8 @@ REM MAKEFILE for 8/16K ROMs using SDCC compiler
 REM CFILENAME = Source name without extension
 REM ROMFILENAME = ROM file name without extension
 REM ROMSIZE = 2000 for 8KBytes ROM or 4000 for 16KBytes ROM
-SET CFILENAME=Test_VDP
-SET ROMFILENAME=TESTVDP
+SET CFILENAME=TestTMS
+SET ROMFILENAME=TESTTMS
 SET CODELOC=0x4020
 SET ROMSIZE=4000
 SET CRT=libs\crt_MSX816kROM4000.rel
