+++ examples/Example02/libs/VDP_TMS9918A_MSXBIOS.rel
@@ -1,23 +1,25 @@
 XL3
-H 9 areas E global symbols
+H 9 areas 10 global symbols
 M VDP_TMS9918A_MSXBIOS
 O -mz80
 S .__.ABS. Def000000
-A _CODE size E0 flags 0 addr 0
-S _CLS Def00005B
-S _CopyFromVRAM Def0000B7
-S _SetSpritesSize Def000019
-S _VPEEK Def000089
-S _SetSpritesZoom Def00002C
-S _FillVRAM Def00008C
-S _COLOR Def00003E
+A _CODE size 116 flags 0 addr 0
+S _CLS Def00002E
+S _CopyFromVRAM Def00008A
+S _SetSpritesSize Def0000BA
+S _SetSpritesZoom Def0000CA
+S _VPEEK Def00005C
+S _FillVRAM Def00005F
+S _COLOR Def000011
 S _SCREEN Def000000
-S _VPOKE Def000076
-S _GetVDP Def0000CD
-S _ClearSprites Def000011
-S _CopyToVRAM Def0000A1
-S _SetVDP Def0000DA
-A _DATA size 0 flags 0 addr 0
+S _VPOKE Def000049
+S _GetVDP Def0000A0
+S _PUTSPRITE Def0000DC
+S _ClearSprites Def0000B2
+S _CopyToVRAM Def000074
+S _SetVDP Def0000AD
+A _DATA size 4 flags 0 addr 0
+S _SPRITE_BF Def000000
 A _INITIALIZED size 0 flags 0 addr 0
 A _DABS size 0 flags 8 addr 0
 A _HOME size 0 flags 0 addr 0
@@ -25,6 +27,10 @@ A _GSINIT size 0 flags 0 addr 0
 A _GSFINAL size 0 flags 0 addr 0
 A _INITIALIZER size 0 flags 0 addr 0
 A _CABS size 0 flags 8 addr 0
+T 00 00 00
+R 00 00 01 00
+T 00 00 00
+R 00 00 01 00
 T 00 00 00
 R 00 00 00 00
 T 00 00 00 DD E5 B7 20 06 3E 28 32 AE F3 AF
@@ -35,83 +41,93 @@ T 0B 00 00 CD 5F 00 DD E1 C9
 R 00 00 00 00
 T 11 00 00
 R 00 00 00 00
-T 11 00 00 DD E5 CD 69 00 DD E1 C9
+T 11 00 00 DD E5 DD 21 00 00 DD 39 32 E9 F3 7D 32
 R 00 00 00 00
-T 19 00 00
+T 1E 00 00 EA F3 DD 7E 04 32 EB F3 CD 62 00 DD E1
 R 00 00 00 00
-T 19 00 00 21 E0 F3 46 FE 01 20 04 CB C8 18 02
+T 2B 00 00 E1 33 E9
 R 00 00 00 00
-T 25 00 00
+T 2E 00 00
 R 00 00 00 00
-T 25 00 00 CB 88
+T 2E 00 00 3A E0 F3 CB 67 28 0A AF 21 00 00 01
 R 00 00 00 00
-T 27 00 00
+T 3A 00 00 C0 03 C3 56 00
 R 00 00 00 00
-T 27 00 00 0E 01 C3 47 00
+T 3F 00 00
 R 00 00 00 00
-T 2C 00 00
+T 3F 00 00 AF 21 00 18 01 00 03 C3 56 00
 R 00 00 00 00
-T 2C 00 00 21 E0 F3 46 B7 28 04 CB C0 18 02
+T 49 00 00
 R 00 00 00 00
-T 37 00 00
+T 49 00 00 DD E5 DD 21 00 00 DD 39 DD 7E 04 CD
 R 00 00 00 00
-T 37 00 00 CB 80
+T 55 00 00 4D 00 DD E1 E1 33 E9
 R 00 00 00 00
-T 39 00 00
+T 5C 00 00
 R 00 00 00 00
-T 39 00 00 0E 01 C3 47 00
+T 5C 00 00 C3 4A 00
 R 00 00 00 00
-T 3E 00 00
+T 5F 00 00
 R 00 00 00 00
-T 3E 00 00 DD E5 DD 21 00 00 DD 39 32 E9 F3 7D 32
+T 5F 00 00 DD E5 DD 21 00 00 DD 39 4B 42 DD 7E 04
 R 00 00 00 00
-T 4B 00 00 EA F3 DD 7E 04 32 EB F3 CD 62 00 DD E1
+T 6C 00 00 CD 56 00 DD E1 E1 33 E9
 R 00 00 00 00
-T 58 00 00 E1 33 E9
+T 74 00 00
 R 00 00 00 00
-T 5B 00 00
+T 74 00 00 DD E5 DD 21 00 00 DD 39 DD 4E 04 DD 46
 R 00 00 00 00
-T 5B 00 00 3A E0 F3 CB 67 28 0A AF 21 00 00 01
+T 81 00 00 05 CD 5C 00 DD E1 E1 F1 E9
 R 00 00 00 00
-T 67 00 00 C0 03 C3 56 00
+T 8A 00 00
 R 00 00 00 00
-T 6C 00 00
+T 8A 00 00 DD E5 DD 21 00 00 DD 39 DD 4E 04 DD 46
 R 00 00 00 00
-T 6C 00 00 AF 21 00 18 01 00 03 C3 56 00
+T 97 00 00 05 CD 59 00 DD E1 E1 F1 E9
 R 00 00 00 00
-T 76 00 00
+T A0 00 00
 R 00 00 00 00
-T 76 00 00 DD E5 DD 21 00 00 DD 39 DD 7E 04 CD
+T A0 00 00 FD 21 DF F3 5F 16 00 FD 19 FD 7E 00 C9
 R 00 00 00 00
-T 82 00 00 4D 00 DD E1 E1 33 E9
+T AD 00 00
 R 00 00 00 00
-T 89 00 00
+T AD 00 00 4F 45 C3 47 00
 R 00 00 00 00
-T 89 00 00 C3 4A 00
+T B2 00 00
 R 00 00 00 00
-T 8C 00 00
+T B2 00 00 DD E5 CD 69 00 DD E1 C9
 R 00 00 00 00
-T 8C 00 00 DD E5 DD 21 00 00 DD 39 4B 42 DD 7E 04
+T BA 00 00
 R 00 00 00 00
-T 99 00 00 CD 56 00 DD E1 E1 33 E9
+T BA 00 00 21 E0 F3 46 FE 01 20 04 CB C8 18 11
 R 00 00 00 00
-T A1 00 00
+T C6 00 00
 R 00 00 00 00
-T A1 00 00 DD E5 DD 21 00 00 DD 39 DD 4E 04 DD 46
+T C6 00 00 CB 88 18 0D
 R 00 00 00 00
-T AE 00 00 05 CD 5C 00 DD E1 E1 F1 E9
+T CA 00 00
 R 00 00 00 00
-T B7 00 00
+T CA 00 00 21 E0 F3 46 B7 28 04 CB C0 18 02
 R 00 00 00 00
-T B7 00 00 DD E5 DD 21 00 00 DD 39 DD 4E 04 DD 46
+T D5 00 00
 R 00 00 00 00
-T C4 00 00 05 CD 59 00 DD E1 E1 F1 E9
+T D5 00 00 CB 80
 R 00 00 00 00
-T CD 00 00
+T D7 00 00
 R 00 00 00 00
-T CD 00 00 FD 21 DF F3 5F 16 00 FD 19 FD 7E 00 C9
+T D7 00 00 0E 01 C3 47 00
 R 00 00 00 00
-T DA 00 00
+T DC 00 00
 R 00 00 00 00
-T DA 00 00 4F 45 CD 47 00 C9
+T DC 00 00 DD E5 DD 21 00 00 DD 39 5F 55 21 00 00
+R 00 00 00 00 00 0E 01 00
+T E9 00 00 DD 4E 04 71 23 72 23 DD 4E 06 CD 8A 00
+R 00 00 00 00
+T F6 00 00 30 04 CB 21 CB 21
+R 00 00 00 00
+T FC 00 00
+R 00 00 00 00
+T FC 00 00 71 23 DD 4E 05 71 7B CD 87 00 11 00 00
+R 00 00 00 00 00 0E 01 00
+T 09 01 00 EB 01 04 00 CD 5C 00 DD E1 E1 F1 33 E9
 R 00 00 00 00
