+++ examples/Example02/source/TestTMS.c
@@ -1,0 +1,854 @@
+/* =============================================================================
+# TestTMS
+
+- Version: 1.3 (1/12/2023)
+- Author: mvac7/303bcn
+- Architecture: MSX
+- Format: 16K ROM
+- Programming language: C and Z80 assembler
+- Compiler: SDCC 4.4 or newer
+
+## Description:
+Perform a test of the functions of the VDP TMS9918A library for MSX ROM 
+environment.
+
+## History of versions:
+- v1.3 ( 1/12/2023) update to SDCC (4.3). For v1.5 library
+- v1.2 (13/ 4/2018)
+============================================================================= */
+
+// ---------------------------------------------------------------------------- Includes
+#include "../include/newTypes.h"
+#include "../include/msxSystemVariables.h"
+#include "../include/msxBIOS.h"
+
+#include "../include/VDP_TMS9918A_MSXBIOS.h"
+
+
+
+#define  HALT __asm halt __endasm   //wait for the next interrupt
+
+
+
+// ---------------------------------------------------------------------------- Labels
+
+
+
+
+// ---------------------------------------------------------------------------- Function Declaration
+void WAIT(uint cicles);
+char INKEY(void);
+void LOCATE(char x, char y);
+
+char isText1Mode(void);
+
+void VPRINT(char column, char line, char* text);  //print in screen 1 or 2
+//void VPOKEARRAY(uint vaddr, char* text);
+
+char PEEK(uint address);
+void POKE(uint address, char value);
+
+void PressKey(void);
+
+
+void TestVDP(void);
+
+void TestClear(char nscr);
+
+void testSCREEN0(void);
+void testSCREEN1(void);
+void testSCREEN2(void);
+void testSCREEN3(void);
+
+void testFill(void);
+void testColor(void);
+void testVpeekVpoke(void);
+
+void showTestCard(void);
+
+void fillOrdered(void);
+
+
+
+void setSpritesPatterns(void);
+void showSprites(char offset);
+void initSprites(void);
+
+void testSPRITES(void);
+void testSpritePosition(void);
+void testSpriteColor(void);
+void testSpritePattern(void);
+void testSpriteVisible(void);
+
+
+
+
+// ---------------------------------------------------------------------------- Constants
+const char text00[] = "MSX fR3eL SDCC Libraries Prj";
+const char text01[] = "Test TMS9918A MSXBIOS Lib"; 
+
+const char msg_PressKey[] = "Press any key to continue";
+
+
+// Project: tMSgFX_def_tilesets
+// tileset pattern
+// Tiles range: 32 to 127
+// Size= 768
+const char TILESET_B0[]={
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0C,0x0C,0x0C,0x0C,0x00,0x0C,0x00,
+0x00,0x14,0x14,0x00,0x00,0x00,0x00,0x00,0x00,0x14,0x3E,0x14,0x3E,0x14,0x00,0x00,
+0x08,0x3E,0x68,0x3E,0x0B,0x3E,0x08,0x00,0x00,0x33,0x36,0x0C,0x1B,0x33,0x00,0x00,
+0x00,0x38,0x6C,0x3D,0x66,0x3D,0x00,0x00,0x00,0x18,0x30,0x00,0x00,0x00,0x00,0x00,
+0x06,0x0C,0x0C,0x0C,0x0C,0x0C,0x06,0x00,0x30,0x18,0x18,0x18,0x18,0x18,0x30,0x00,
+0x00,0x5A,0x3C,0x7E,0x3C,0x5A,0x00,0x00,0x00,0x18,0x18,0x7E,0x18,0x18,0x00,0x00,
+0x00,0x00,0x00,0x00,0x18,0x18,0x10,0x00,0x00,0x00,0x00,0x7E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x06,0x0C,0x18,0x30,0x60,0x00,0x00,
+0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x18,0x38,0x18,0x18,0x3C,0x00,0x00,
+0x00,0x7E,0x03,0x3E,0x60,0x7F,0x00,0x00,0x00,0x7E,0x03,0x1E,0x03,0x7E,0x00,0x00,
+0x00,0x63,0x63,0x7F,0x03,0x03,0x00,0x00,0x00,0x7F,0x60,0x7E,0x03,0x7E,0x00,0x00,
+0x00,0x3E,0x60,0x7E,0x63,0x3E,0x00,0x00,0x00,0x7F,0x03,0x06,0x0C,0x0C,0x00,0x00,
+0x00,0x3E,0x63,0x3E,0x63,0x3E,0x00,0x00,0x00,0x3E,0x63,0x3F,0x03,0x3E,0x00,0x00,
+0x00,0x18,0x18,0x00,0x18,0x18,0x00,0x00,0x00,0x18,0x18,0x00,0x18,0x18,0x10,0x00,
+0x00,0x0C,0x18,0x30,0x18,0x0C,0x00,0x00,0x00,0x00,0x3C,0x00,0x3C,0x00,0x00,0x00,
+0x00,0x30,0x18,0x0C,0x18,0x30,0x00,0x00,0x00,0x1E,0x33,0x06,0x0C,0x00,0x0C,0x00,
+0x3E,0x41,0x5D,0x55,0x5F,0x40,0x3E,0x00,0x00,0x3E,0x63,0x7F,0x63,0x63,0x00,0x00,
+0x00,0x7E,0x63,0x7E,0x63,0x7E,0x00,0x00,0x00,0x3F,0x60,0x60,0x60,0x3F,0x00,0x00,
+0x00,0x7E,0x63,0x63,0x63,0x7E,0x00,0x00,0x00,0x7F,0x60,0x7C,0x60,0x7F,0x00,0x00,
+0x00,0x7F,0x60,0x7C,0x60,0x60,0x00,0x00,0x00,0x3F,0x60,0x67,0x63,0x3F,0x00,0x00,
+0x00,0x63,0x63,0x7F,0x63,0x63,0x00,0x00,0x00,0x0C,0x0C,0x0C,0x0C,0x0C,0x00,0x00,
+0x00,0x0C,0x0C,0x0C,0x0C,0x38,0x00,0x00,0x00,0x63,0x66,0x7C,0x66,0x63,0x00,0x00,
+0x00,0x60,0x60,0x60,0x60,0x7F,0x00,0x00,0x00,0x63,0x77,0x7F,0x6B,0x63,0x00,0x00,
+0x00,0x63,0x73,0x7B,0x6F,0x67,0x00,0x00,0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,
+0x00,0x7E,0x63,0x63,0x7E,0x60,0x00,0x00,0x00,0x3E,0x63,0x6B,0x66,0x3D,0x00,0x00,
+0x00,0x7E,0x63,0x63,0x7E,0x63,0x00,0x00,0x00,0x3F,0x60,0x3E,0x03,0x7E,0x00,0x00,
+0x00,0x7E,0x18,0x18,0x18,0x18,0x00,0x00,0x00,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,
+0x00,0x63,0x63,0x63,0x36,0x1C,0x00,0x00,0x00,0x63,0x6B,0x7F,0x77,0x63,0x00,0x00,
+0x00,0x63,0x36,0x1C,0x36,0x63,0x00,0x00,0x00,0x66,0x66,0x3C,0x18,0x18,0x00,0x00,
+0x00,0x7E,0x0C,0x18,0x30,0x7E,0x00,0x00,0x0E,0x0C,0x0C,0x0C,0x0C,0x0C,0x0E,0x00,
+0x00,0x60,0x30,0x18,0x0C,0x06,0x00,0x00,0x38,0x18,0x18,0x18,0x18,0x18,0x38,0x00,
+0x10,0x38,0x6C,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7E,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3E,0x03,0x3F,0x63,0x7F,0x00,0x00,
+0x60,0x7E,0x63,0x63,0x63,0x7E,0x00,0x00,0x00,0x3F,0x60,0x60,0x60,0x3F,0x00,0x00,
+0x03,0x3F,0x63,0x63,0x63,0x3F,0x00,0x00,0x00,0x3E,0x63,0x7F,0x60,0x3E,0x00,0x00,
+0x00,0x3E,0x60,0x78,0x60,0x60,0x00,0x00,0x00,0x3E,0x63,0x63,0x3F,0x03,0x3E,0x00,
+0x60,0x7E,0x63,0x63,0x63,0x63,0x00,0x00,0x0C,0x00,0x0C,0x0C,0x0C,0x0C,0x00,0x00,
+0x0C,0x00,0x0C,0x0C,0x0C,0x0C,0x38,0x00,0x00,0x60,0x66,0x7C,0x66,0x63,0x00,0x00,
+0x18,0x18,0x18,0x18,0x18,0x18,0x00,0x00,0x00,0x7E,0x6B,0x6B,0x6B,0x6B,0x00,0x00,
+0x00,0x7E,0x63,0x63,0x63,0x63,0x00,0x00,0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,
+0x00,0x7E,0x63,0x63,0x63,0x7E,0x60,0x00,0x00,0x3F,0x63,0x63,0x63,0x3F,0x03,0x00,
+0x00,0x3F,0x60,0x60,0x60,0x60,0x00,0x00,0x00,0x3F,0x60,0x3E,0x03,0x7E,0x00,0x00,
+0x18,0x3E,0x18,0x18,0x18,0x0E,0x00,0x00,0x00,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,
+0x00,0x63,0x63,0x63,0x36,0x1C,0x00,0x00,0x00,0x63,0x6B,0x6B,0x6B,0x36,0x00,0x00,
+0x00,0x66,0x3C,0x18,0x3C,0x66,0x00,0x00,0x00,0x63,0x63,0x63,0x3F,0x03,0x3E,0x00,
+0x00,0x7E,0x0C,0x18,0x30,0x7E,0x00,0x00,0x06,0x0C,0x0C,0x18,0x0C,0x0C,0x06,0x00,
+0x18,0x18,0x18,0x00,0x18,0x18,0x18,0x00,0x30,0x18,0x18,0x0C,0x18,0x18,0x30,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00};
+
+
+
+const char sprcol[8]={12,2,3,7,6,8,9,14};
+
+
+
+const char SPRITE_DATA[]={
+60,66,165,129,165,153,66,60,
+60,126,219,255,255,219,102,60,
+108,254,254,254,124,56,16,0,
+16,56,124,254,124,56,16,0,
+16,56,84,254,84,16,56,0,
+16,56,124,254,254,16,56,0,
+0,0,0,24,24,0,0,0,
+255,255,255,231,231,255,255,255,
+15,31,61,63,123,124,191,191,159,239,111,31,15,5,5,29,
+192,224,96,192,64,128,128,132,134,222,192,192,128,0,0,192,
+5,7,26,46,111,91,92,91,95,88,111,45,22,2,2,6,
+64,192,184,212,212,106,234,106,234,106,212,212,248,64,64,96,
+3,7,79,207,77,79,44,30,63,63,63,63,31,2,2,6,
+192,226,230,226,162,244,56,124,254,254,254,252,248,64,64,96,
+5,30,63,111,87,110,63,0,7,5,7,4,6,3,2,6,
+96,248,244,234,118,190,124,0,224,96,224,32,96,192,64,96,
+5,7,31,63,127,127,127,125,127,127,120,63,31,4,4,12,
+64,192,240,248,252,252,252,124,252,252,60,248,240,64,64,96,
+7,31,63,63,127,113,127,127,127,126,62,63,31,7,2,6,
+224,248,252,252,254,142,254,254,254,126,124,252,248,224,64,96,
+0,7,31,63,127,113,255,255,123,124,63,31,7,2,2,6,
+0,224,248,252,254,142,255,255,222,62,252,248,224,64,64,96,
+5,30,63,111,87,110,63,0,7,5,7,4,6,3,2,6,
+96,248,244,234,118,190,124,0,224,96,224,32,96,192,64,96
+};
+
+
+  
+// Sine
+// Length=255; Min=16; Max=144; Freq=1; Phase=0
+const char SIN[]={
+0x50,0x52,0x53,0x55,0x56,0x58,0x59,0x5B,0x5D,0x5E,0x60,0x61,0x63,0x64,0x66,0x67,
+0x69,0x6A,0x6B,0x6D,0x6E,0x70,0x71,0x72,0x74,0x75,0x76,0x78,0x79,0x7A,0x7B,0x7C,
+0x7D,0x7E,0x80,0x81,0x82,0x83,0x84,0x84,0x85,0x86,0x87,0x88,0x89,0x89,0x8A,0x8B,
+0x8B,0x8C,0x8C,0x8D,0x8D,0x8E,0x8E,0x8F,0x8F,0x8F,0x8F,0x90,0x90,0x90,0x90,0x90,
+0x90,0x90,0x90,0x90,0x90,0x8F,0x8F,0x8F,0x8F,0x8E,0x8E,0x8E,0x8D,0x8D,0x8C,0x8C,
+0x8B,0x8A,0x8A,0x89,0x88,0x87,0x87,0x86,0x85,0x84,0x83,0x82,0x81,0x80,0x7F,0x7E,
+0x7D,0x7C,0x7B,0x79,0x78,0x77,0x76,0x74,0x73,0x72,0x70,0x6F,0x6E,0x6C,0x6B,0x69,
+0x68,0x66,0x65,0x63,0x62,0x60,0x5F,0x5D,0x5C,0x5A,0x59,0x57,0x56,0x54,0x52,0x51,
+0x4F,0x4E,0x4C,0x4A,0x49,0x47,0x46,0x44,0x43,0x41,0x40,0x3E,0x3D,0x3B,0x3A,0x38,
+0x37,0x35,0x34,0x32,0x31,0x30,0x2E,0x2D,0x2C,0x2A,0x29,0x28,0x27,0x25,0x24,0x23,
+0x22,0x21,0x20,0x1F,0x1E,0x1D,0x1C,0x1B,0x1A,0x19,0x19,0x18,0x17,0x16,0x16,0x15,
+0x14,0x14,0x13,0x13,0x12,0x12,0x12,0x11,0x11,0x11,0x11,0x10,0x10,0x10,0x10,0x10,
+0x10,0x10,0x10,0x10,0x10,0x11,0x11,0x11,0x11,0x12,0x12,0x13,0x13,0x14,0x14,0x15,
+0x15,0x16,0x17,0x17,0x18,0x19,0x1A,0x1B,0x1C,0x1C,0x1D,0x1E,0x1F,0x20,0x22,0x23,
+0x24,0x25,0x26,0x27,0x28,0x2A,0x2B,0x2C,0x2E,0x2F,0x30,0x32,0x33,0x35,0x36,0x37,
+0x39,0x3A,0x3C,0x3D,0x3F,0x40,0x42,0x43,0x45,0x47,0x48,0x4A,0x4B,0x4D,0x4E,0x50};
+
+
+
+// ---------------------------------------------------------------------------- Global Variables
+char spr_posX[8];
+char spr_posY[8];
+
+
+
+// ---------------------------------------------------------------------------- Functions
+
+
+//
+void main(void)
+{
+	COLOR(15,4,5);
+	POKE(LINL32,32);
+	SCREEN(1);
+
+	VPRINT(0,0,text00);
+	VPRINT(0,1,text01);
+		
+/*	VPRINT(0,7,msg_PressKey);
+	LOCATE(25,7);
+	INKEY();*/
+	PressKey();
+	
+  
+//TEST -------------------------------------------------------------------------   
+	TestVDP();
+	
+//	TestCLS();
+//	WAIT(100);
+	
+	testColor();
+	WAIT(100);
+
+	testSCREEN0();  
+	WAIT(100);
+
+	testSCREEN1();  
+	WAIT(100);
+
+	testSCREEN2();  
+	WAIT(100);
+
+	testSCREEN3();  
+	WAIT(100);	
+
+	testSPRITES();
+	WAIT(100);	
+
+//END --------------------------------------------------------------------------  
+  
+	COLOR(15,4,4);
+	SCREEN(1);
+
+	VPRINT(0,0,"END");
+	WAIT(30*10);
+   
+}
+
+
+
+/* =============================================================================
+INKEY
+Description: 
+		One character input (waiting)
+Input:	-       
+Output:	-		
+============================================================================= */
+char INKEY(void) __naked
+{
+__asm 
+  
+	jp BIOS_CHGET
+
+__endasm;
+}
+
+
+
+/* =============================================================================
+LOCATE
+
+Description: 
+		Moves the cursor to the specified location.
+Input:	(char) Position X of the cursor. (0 to 31 or 79)
+		(char) Position Y of the cursor. (0 to 23)         
+Output:	-
+============================================================================= */
+void LOCATE(char x, char y) __naked
+{
+x;
+y;
+__asm
+
+  inc  A
+  ld   H,A
+
+  inc  L
+  
+  jp BIOS_POSIT
+
+__endasm;
+}
+
+
+
+/* =============================================================================
+   Indicates whether Text 1 mode is active.
+   Output:	1=Yes/True ; 0=No/False
+============================================================================= */
+char isText1Mode(void)
+{
+	//char A = *(unsigned int *) 0xF3E0;	//RG1SAV
+	//if (A&0b00010000) return 1; 		//Text 40col Mode
+	if (GetVDP(1)&0b00010000) return 1; 		//Text 40col Mode
+	return 0;	
+}
+
+
+
+//print in screen 1 or 2
+void VPRINT(char column, char line, char* text)
+{
+	uint vaddr;
+	if (GetVDP(1)&0b00010000) vaddr=BASE0 + (line*40)+column;
+	else vaddr=BASE10 + (line*32)+column; // calcula la posicion en la VRAM
+	//VPOKEARRAY(vaddr, text);
+	while(*(text)) VPOKE(vaddr++,*(text++));
+}
+
+
+
+/*void VPOKEARRAY(uint vaddr, char* text)
+{
+  while(*(text)) VPOKE(vaddr++,*(text++));
+}*/
+
+
+
+
+// Generates a pause in the execution of n interruptions.
+// PAL: 50=1second. ; NTSC: 60=1second. 
+void WAIT(uint cicles)
+{
+  uint i;
+  for(i=0;i<cicles;i++) HALT;
+}
+
+
+
+char PEEK(uint address) __naked
+{
+address;
+__asm
+
+  ld   A,(HL)
+
+  ret
+__endasm;
+}
+
+
+
+void POKE(uint address,char value)
+{
+address;value;
+__asm
+  push IX
+  ld   IX,#0
+  add  IX,SP
+  
+  ld   A,4(IX)
+  ld   (HL),A
+
+  pop  IX
+__endasm;
+}
+
+
+
+void PressKey(void)
+{
+	VPRINT(0,23,msg_PressKey);
+	LOCATE(25,23);
+	INKEY();	
+}
+
+
+
+void TestVDP(void)
+{
+	char VDPval_before;
+	char VDPval_after;
+	
+	SCREEN(1);
+	VDPval_before = GetVDP(VDP_Mode1);
+	
+	VPRINT(0,0,">Test SetVDP()");
+	WAIT(100);
+	SetVDP(VDP_Mode1,VDPval_before|0b00010000);	//set M1=1
+	HALT;
+	VDPval_after = GetVDP(VDP_Mode1);
+	WAIT(50);
+	SetVDP(VDP_Mode1,VDPval_before);	//restore screen
+	HALT;
+	VPRINT(0,1,">Test GetVDP()");
+	WAIT(50);
+	if (VDPval_after&0b00010000) VPRINT(0,2,">Test=Ok");	//VDPval_before != Textmode1 and VDPval_after == Textmode1
+	else VPRINT(0,2,"Test=ERROR!");
+		
+	PressKey();	
+}
+
+
+	
+void TestClear(char nscr)
+{	
+	unsigned int BC;
+	unsigned int vaddr;	
+	unsigned int vsize;
+	boolean testResult=true;
+
+	if(nscr==0){
+		vaddr=T1_MAP;
+		vsize=0x3C0;
+	}else{
+		vaddr=G1_MAP;
+		vsize=0x300;
+	}
+			
+
+	WAIT(100);
+	
+	CLS();
+	WAIT(25);
+	
+	for(BC=0;BC<vsize;BC++) if(VPEEK(vaddr++)!=0) testResult=false;
+		
+	WAIT(25);
+	if(testResult==true) VPRINT(0,0,">Test CLS()=Ok");
+	else VPRINT(0,0,">Test CLS()=ERROR!");
+	
+	PressKey();
+}
+
+
+
+// TEST SCREEN 0 ###############################################################
+void testSCREEN0(void)
+{
+	char i;
+
+	COLOR(LIGHT_GREEN,DARK_GREEN,DARK_GREEN);    
+	SCREEN(0);
+
+	VPRINT(31,23," SCREEN 0");
+	WAIT(50);
+
+	//fillOrdered();
+	for(i=0;i<255;i++) VPOKE(BASE0+i,i);
+	WAIT(50);
+
+	testVpeekVpoke();
+	WAIT(50);
+
+	//testFill
+	VPRINT(0,0,"FillVRAM()");
+	for(i=32;i<95;i++)
+	{
+		WAIT(4);
+
+		FillVRAM (BASE0+10, 942, i);    
+	}
+	
+	FillVRAM(T1_MAP, 0x3C0, 9);	//character 9=circle
+	VPRINT(0,0,">Test CLS() SCREEN 0");
+	TestClear(0);
+  
+}
+
+
+
+// TEST SCREEN 1 ###############################################################
+void testSCREEN1(void)
+{
+	COLOR(15,5,1);    
+	SCREEN(1);
+
+	VPRINT(23,23," SCREEN 1");
+	WAIT(50);
+
+	fillOrdered();
+	WAIT(50);
+
+	VPRINT(0,10,"Test CopyToVRAM()");  
+	WAIT(50);
+
+	//copy to VRAM tileset, only gfx patterns
+	CopyToVRAM((uint) TILESET_B0,BASE12+(32*8),96*8);
+	WAIT(50);
+
+	/* colors
+	The BG Colors array defines 32 colors (each 4 bit background, and 4 bit 
+	foreground, as in VDP register 7). The colors are assigned to the BG Tiles as
+	follows: Tiles 00..07 share the first color, tiles 08..0F share the second 
+	color, etc, and tiles F8..FF share the last color.*/
+	FillVRAM(BASE6,32,0x75);
+
+	WAIT(50);
+
+	testFill();
+	
+	FillVRAM(G1_MAP, 0x300, 203);
+	VPRINT(0,0,">Test CLS() SCREEN 1");
+	TestClear(1);
+  
+}
+
+
+
+// TEST SCREEN 2 ###############################################################
+void testSCREEN2(void)
+{
+	COLOR(15,4,1);
+	SCREEN(2);
+
+	VPRINT(23,23," SCREEN 2");
+
+	//copy to VRAM tileset, only gfx patterns and fill colors
+	//bank 0
+	HALT;
+	CopyToVRAM((uint) TILESET_B0,BASE12+(32*8),96*8);
+	FillVRAM(BASE11+(32*8),96*8,0x96); //colors (Red)
+
+	//bank 1
+	CopyToVRAM((uint) TILESET_B0,BASE12+0x800+(32*8),96*8);
+	FillVRAM(BASE11+BANK1+(32*8),96*8,0x3C); //colors (green)
+
+	//bank 2
+	CopyToVRAM((uint) TILESET_B0,BASE12+0x1000+(32*8),96*8);
+	FillVRAM(BASE11+BANK2+(32*8),96*8,0x54); //colors (blue)
+	WAIT(50);
+	//END copy
+
+
+	//test CopyFromVRAM
+	VPRINT(0,5,"Test CopyFromVRAM()");
+	//copy VRAM to RAM
+	CopyFromVRAM(BASE12+(32*8),0xE000,64*8);
+	WAIT(50);
+
+	//copy RAM to VRAM
+	CopyToVRAM(0xE000,BASE12+0x600,64*8);   
+	FillVRAM(BASE11+0x600,64*8,0xFE); //colors
+	WAIT(120);
+	//END CopyFromVRAM
+
+
+	//test fill VRAM  
+	testFill();
+	
+	FillVRAM(G2_MAP, 0x300, '#');
+	VPRINT(0,0,">Test CLS() SCREEN 2");
+	TestClear(2);
+
+}
+
+
+
+// TEST SCREEN 3 ###############################################################
+void testSCREEN3(void)
+{
+	char value=0;
+	uint i;
+	uint vaddr = BASE17;
+
+	COLOR(15,4,4);
+	SCREEN(1);
+
+	VPRINT(0,0,"Test SCREEN 3");  
+	WAIT(200);
+	 
+	SCREEN(3);
+
+	for(i=0;i<0x600;i++)
+	{
+		VPOKE(vaddr++,value);
+		if(value<255) value++;
+		else value=0; 
+	}
+	WAIT(250);
+
+}
+
+
+
+// show tileset
+void fillOrdered(void)
+{
+	char i;
+	for(i=0;i<255;i++) VPOKE(BASE5+i,i);
+}
+
+
+
+// TEST COLOR  #################################################################
+void testColor(void)
+{
+	char i;
+	char inkcol=15;
+
+	COLOR(15,0,4);
+	SCREEN(0);
+
+	VPRINT(15,10,"TEST color");
+	WAIT(50);
+
+	for(i=0;i<16;i++)
+	{ 
+		COLOR(inkcol--,i,i);
+			
+		WAIT(15);
+	}  
+}
+
+
+
+// TEST VPEEK & VPOKE  #########################################################
+// copy a bloq using vpeek and vpoke
+void testVpeekVpoke(void)
+{
+	uint vaddr = BASE0;
+	char i;
+	char value;
+
+	VPRINT(0,9,"TEST VPEEK & VPOKE:");
+
+	for(i=0;i<255;i++)
+	{
+		HALT;
+		value = VPEEK(vaddr);
+		VPOKE(vaddr+440,value);
+		vaddr++;  
+	}  
+
+}
+
+
+
+// TEST FILL  ##################################################################
+//fill screen map with a characters range 
+void testFill(void)
+{
+	char i;
+
+	VPRINT(0,0,"FillVRAM()");
+	for(i=32;i<95;i++)
+	{
+		HALT;
+
+		FillVRAM (BASE10+10, 750, i);
+
+		HALT;
+		HALT;
+		HALT;
+	}  
+  
+}
+
+
+
+// TEST SPRITES  ###############################################################
+void testSPRITES(void)
+{
+	char posY=0;
+	
+	COLOR(15,4,5);
+	SCREEN(1);
+
+	initSprites();
+	
+	CopyToVRAM((uint) TILESET_B0,BASE12+(32*8),96*8);
+	setSpritesPatterns();	
+
+	VPRINT(0,posY++,"Test SPRITEs -----------------"); 
+	posY++;
+
+	// sprites 8x8
+	VPRINT(0,posY++, "SetSpritesSize(0) SPRITES 8x8");
+	SetSpritesSize(0);
+	SetSpritesZoom(false);  
+	showSprites(0);
+	WAIT(50);
+
+	// test sprites 16x16
+	VPRINT(0,posY++, "SetSpritesSize(1) SPRITES 16x16");
+	SetSpritesSize(1);
+	showSprites(2);
+	WAIT(50);
+
+	// test sprites 16x16 + zoom  
+	VPRINT(0,posY++, "SetSpritesZoom(true) SPRITES x2");
+	SetSpritesZoom(true);
+	WAIT(50);
+
+	//test clear sprites data
+	VPRINT(0,posY++, "ClearSprites()");
+	ClearSprites();
+	WAIT(100);
+
+	setSpritesPatterns();
+	VPRINT(0,posY++, "PUTSPRITE(plane,x,y,color,nSPR)");
+	showSprites(2);
+	WAIT(50);	
+
+	VPRINT(0,posY++, "Visible");
+	testSpriteVisible();
+
+	VPRINT(0,posY++, "Change Pattern"); 
+	testSpritePattern();
+
+	VPRINT(0,posY++, "Change Color"); 
+	testSpriteColor();
+
+	VPRINT(0,posY++, "Change Position");
+	testSpritePosition();
+
+	WAIT(50);
+
+	VPRINT(0,posY++, "Set EarlyClock");
+	PUTSPRITE(7, spr_posX[7], spr_posY[7], 8+127, 9);
+	WAIT(50);
+
+	VPRINT(0,posY++, "Unset EarlyClock");
+	PUTSPRITE(7, spr_posX[7], spr_posY[7], 8, 9);
+//	WAIT(50);
+
+/*	posY+=3;
+	VPRINT(0,posY, "Press any key.");
+	LOCATE(12,posY);
+	INKEY();*/
+	
+	PressKey();
+}
+
+
+
+
+// Copy sprites data from memory to VRAM
+void setSpritesPatterns(void)
+{
+	HALT;
+	CopyToVRAM((uint) SPRITE_DATA,BASE14,32*10);
+}
+
+
+
+void initSprites(void)
+{
+	char X=0,Y=3;
+	char i=0;
+
+	for(i=0;i<8;i++)
+	{
+		spr_posX[i]=X*32;
+		spr_posY[i]=Y*32;
+		X++;
+		if(X==4)
+		{
+			X=0;
+			Y++;
+		}
+	}
+}
+
+
+
+// TEST PUTSPRITE  #############################################################
+void showSprites(char offset)
+{
+	char X=0,Y=3;
+	char i=0;
+
+	for(i=0;i<8;i++) PUTSPRITE(i, spr_posX[i], spr_posY[i], sprcol[i], i+offset);
+
+}
+
+
+
+// TEST SETSPRITEVISIBLE  ######################################################
+void testSpriteVisible(void)
+{
+	char i,o;
+
+	for(o=0;o<8;o++)
+	{
+		for(i=0;i<8;i++)
+		{
+			if (i==o) PUTSPRITE(i, spr_posX[i], spr_posY[i], sprcol[i], i+2);
+			else PUTSPRITE(i, 0, SPRITES_YHIDDEN, 0, 0);  
+		}
+		WAIT(25);  
+	}  
+}
+
+
+
+// TEST SETSPRITEPATTERN  ######################################################
+void testSpritePattern(void)
+{
+	char i;
+
+	for(i=2;i<10;i++)
+	{
+		PUTSPRITE(7, spr_posX[7], spr_posY[7], sprcol[7], i);
+		WAIT(25);  
+	}  
+}
+
+
+
+// TEST SETSPRITECOLOR  ########################################################
+void testSpriteColor(void)
+{
+	char i;
+
+	for(i=0;i<16;i++)
+	{
+		PUTSPRITE(7, spr_posX[7], spr_posY[7], i, 9);
+		WAIT(25);  
+	}  
+}
+
+
+
+// TEST SETSPRITEPOSITION  #####################################################
+void testSpritePosition(void)
+{
+	uint i=0;
+	char gradX = 64;
+	char gradY = 0;
+
+	for(i=0;i<660;i++)
+	{
+		HALT;
+		PUTSPRITE(7, SIN[gradX], SIN[gradY], 8, 9);
+		gradX++;
+		gradY++;  
+	}
+}
+
