--- examples/Example02/source/Test_VDP.c
@@ -1,605 +1,0 @@
-/* =============================================================================
-Test VDP TMS9918A MSX BIOS Library (fR3eL Project)
-
-Version: 1.3 (1/12/2023)
-Author: mvac7/303bcn
-Architecture: MSX
-Format: 16K ROM
-Programming language: C and Z80 assembler
-Compiler: SDCC 4.4 or newer
-
-Description:
-Perform a test of the functions of the VDP TMS9918A library for MSX ROM 
-environment.
-
-History of versions:
-- v1.3 ( 1/12/2023) update to SDCC (4.3). For v1.5 library
-- v1.2 (13/ 4/2018)
-============================================================================= */
-
-#include "../include/newTypes.h"
-#include "../include/msxSystemVariables.h"
-#include "../include/msxBIOS.h"
-
-#include "../include/VDP_TMS9918A_MSXBIOS.h"
-
-
-
-#define  HALT __asm halt __endasm   //wait for the next interrupt
-
-
-
-// Definitions of Labels -------------------------------------------------------
-
-
-
-
-//  definition of functions  ---------------------------------------------------
-void WAIT(uint cicles);
-char INKEY(void);
-void LOCATE(char x, char y);
-
-void VPRINT(char column, char line, char* text);  //print in screen 1 or 2
-//void VPOKEARRAY(uint vaddr, char* text);
-
-char PEEK(uint address);
-void POKE(uint address, char value);
-
-void PressKey(void);
-
-
-void TestVDP(void);
-
-void TestCLS(void);
-void TestClear(char nscr);
-
-void testSCREEN0(void);
-void testSCREEN1(void);
-void testSCREEN2(void);
-void testSCREEN3(void);
-
-void testFill(void);
-void testColor(void);
-void testVpeekVpoke(void);
-
-void showTestCard(void);
-
-void fillOrdered(void);
-
-char isText1Mode(void);
-
-
-
-// constants  ------------------------------------------------------------------
-const char text00[] = "MSX fR3eL SDCC Libraries Prj";
-const char text01[] = "Test TMS9918A MSXBIOS Lib"; 
-
-const char text_presskey[] = "Press any key to continue";
-
-
-// Project: tMSgFX_def_tilesets
-// tileset pattern
-// Tiles range: 32 to 127
-// Size= 768
-const char TILESET_B0[]={
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x0C,0x0C,0x0C,0x0C,0x00,0x0C,0x00,
-0x00,0x14,0x14,0x00,0x00,0x00,0x00,0x00,0x00,0x14,0x3E,0x14,0x3E,0x14,0x00,0x00,
-0x08,0x3E,0x68,0x3E,0x0B,0x3E,0x08,0x00,0x00,0x33,0x36,0x0C,0x1B,0x33,0x00,0x00,
-0x00,0x38,0x6C,0x3D,0x66,0x3D,0x00,0x00,0x00,0x18,0x30,0x00,0x00,0x00,0x00,0x00,
-0x06,0x0C,0x0C,0x0C,0x0C,0x0C,0x06,0x00,0x30,0x18,0x18,0x18,0x18,0x18,0x30,0x00,
-0x00,0x5A,0x3C,0x7E,0x3C,0x5A,0x00,0x00,0x00,0x18,0x18,0x7E,0x18,0x18,0x00,0x00,
-0x00,0x00,0x00,0x00,0x18,0x18,0x10,0x00,0x00,0x00,0x00,0x7E,0x00,0x00,0x00,0x00,
-0x00,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x06,0x0C,0x18,0x30,0x60,0x00,0x00,
-0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x18,0x38,0x18,0x18,0x3C,0x00,0x00,
-0x00,0x7E,0x03,0x3E,0x60,0x7F,0x00,0x00,0x00,0x7E,0x03,0x1E,0x03,0x7E,0x00,0x00,
-0x00,0x63,0x63,0x7F,0x03,0x03,0x00,0x00,0x00,0x7F,0x60,0x7E,0x03,0x7E,0x00,0x00,
-0x00,0x3E,0x60,0x7E,0x63,0x3E,0x00,0x00,0x00,0x7F,0x03,0x06,0x0C,0x0C,0x00,0x00,
-0x00,0x3E,0x63,0x3E,0x63,0x3E,0x00,0x00,0x00,0x3E,0x63,0x3F,0x03,0x3E,0x00,0x00,
-0x00,0x18,0x18,0x00,0x18,0x18,0x00,0x00,0x00,0x18,0x18,0x00,0x18,0x18,0x10,0x00,
-0x00,0x0C,0x18,0x30,0x18,0x0C,0x00,0x00,0x00,0x00,0x3C,0x00,0x3C,0x00,0x00,0x00,
-0x00,0x30,0x18,0x0C,0x18,0x30,0x00,0x00,0x00,0x1E,0x33,0x06,0x0C,0x00,0x0C,0x00,
-0x3E,0x41,0x5D,0x55,0x5F,0x40,0x3E,0x00,0x00,0x3E,0x63,0x7F,0x63,0x63,0x00,0x00,
-0x00,0x7E,0x63,0x7E,0x63,0x7E,0x00,0x00,0x00,0x3F,0x60,0x60,0x60,0x3F,0x00,0x00,
-0x00,0x7E,0x63,0x63,0x63,0x7E,0x00,0x00,0x00,0x7F,0x60,0x7C,0x60,0x7F,0x00,0x00,
-0x00,0x7F,0x60,0x7C,0x60,0x60,0x00,0x00,0x00,0x3F,0x60,0x67,0x63,0x3F,0x00,0x00,
-0x00,0x63,0x63,0x7F,0x63,0x63,0x00,0x00,0x00,0x0C,0x0C,0x0C,0x0C,0x0C,0x00,0x00,
-0x00,0x0C,0x0C,0x0C,0x0C,0x38,0x00,0x00,0x00,0x63,0x66,0x7C,0x66,0x63,0x00,0x00,
-0x00,0x60,0x60,0x60,0x60,0x7F,0x00,0x00,0x00,0x63,0x77,0x7F,0x6B,0x63,0x00,0x00,
-0x00,0x63,0x73,0x7B,0x6F,0x67,0x00,0x00,0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,
-0x00,0x7E,0x63,0x63,0x7E,0x60,0x00,0x00,0x00,0x3E,0x63,0x6B,0x66,0x3D,0x00,0x00,
-0x00,0x7E,0x63,0x63,0x7E,0x63,0x00,0x00,0x00,0x3F,0x60,0x3E,0x03,0x7E,0x00,0x00,
-0x00,0x7E,0x18,0x18,0x18,0x18,0x00,0x00,0x00,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,
-0x00,0x63,0x63,0x63,0x36,0x1C,0x00,0x00,0x00,0x63,0x6B,0x7F,0x77,0x63,0x00,0x00,
-0x00,0x63,0x36,0x1C,0x36,0x63,0x00,0x00,0x00,0x66,0x66,0x3C,0x18,0x18,0x00,0x00,
-0x00,0x7E,0x0C,0x18,0x30,0x7E,0x00,0x00,0x0E,0x0C,0x0C,0x0C,0x0C,0x0C,0x0E,0x00,
-0x00,0x60,0x30,0x18,0x0C,0x06,0x00,0x00,0x38,0x18,0x18,0x18,0x18,0x18,0x38,0x00,
-0x10,0x38,0x6C,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7E,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x3E,0x03,0x3F,0x63,0x7F,0x00,0x00,
-0x60,0x7E,0x63,0x63,0x63,0x7E,0x00,0x00,0x00,0x3F,0x60,0x60,0x60,0x3F,0x00,0x00,
-0x03,0x3F,0x63,0x63,0x63,0x3F,0x00,0x00,0x00,0x3E,0x63,0x7F,0x60,0x3E,0x00,0x00,
-0x00,0x3E,0x60,0x78,0x60,0x60,0x00,0x00,0x00,0x3E,0x63,0x63,0x3F,0x03,0x3E,0x00,
-0x60,0x7E,0x63,0x63,0x63,0x63,0x00,0x00,0x0C,0x00,0x0C,0x0C,0x0C,0x0C,0x00,0x00,
-0x0C,0x00,0x0C,0x0C,0x0C,0x0C,0x38,0x00,0x00,0x60,0x66,0x7C,0x66,0x63,0x00,0x00,
-0x18,0x18,0x18,0x18,0x18,0x18,0x00,0x00,0x00,0x7E,0x6B,0x6B,0x6B,0x6B,0x00,0x00,
-0x00,0x7E,0x63,0x63,0x63,0x63,0x00,0x00,0x00,0x3E,0x63,0x63,0x63,0x3E,0x00,0x00,
-0x00,0x7E,0x63,0x63,0x63,0x7E,0x60,0x00,0x00,0x3F,0x63,0x63,0x63,0x3F,0x03,0x00,
-0x00,0x3F,0x60,0x60,0x60,0x60,0x00,0x00,0x00,0x3F,0x60,0x3E,0x03,0x7E,0x00,0x00,
-0x18,0x3E,0x18,0x18,0x18,0x0E,0x00,0x00,0x00,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,
-0x00,0x63,0x63,0x63,0x36,0x1C,0x00,0x00,0x00,0x63,0x6B,0x6B,0x6B,0x36,0x00,0x00,
-0x00,0x66,0x3C,0x18,0x3C,0x66,0x00,0x00,0x00,0x63,0x63,0x63,0x3F,0x03,0x3E,0x00,
-0x00,0x7E,0x0C,0x18,0x30,0x7E,0x00,0x00,0x06,0x0C,0x0C,0x18,0x0C,0x0C,0x06,0x00,
-0x18,0x18,0x18,0x00,0x18,0x18,0x18,0x00,0x30,0x18,0x18,0x0C,0x18,0x18,0x30,0x00,
-0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00};
-
-
-
-// Functions -------------------------------------------------------------------
-
-
-//
-void main(void)
-{
-	COLOR(15,4,5);
-	POKE(LINL32,32);
-	SCREEN(1);
-
-	VPRINT(0,0,text00);
-	VPRINT(0,1,text01);
-		
-/*	VPRINT(0,7,text_presskey);
-	LOCATE(25,7);
-	INKEY();*/
-	PressKey();
-	
-	TestVDP();
-  
-//TEST -------------------------------------------------------------------------   
-	TestCLS();
-	WAIT(100);
-	
-	testColor();
-	WAIT(100);
-
-	testSCREEN0();  
-	WAIT(100);
-
-	testSCREEN1();  
-	WAIT(100);
-
-	testSCREEN2();  
-	WAIT(100);
-
-	testSCREEN3();  
-	WAIT(100);
-
-//END --------------------------------------------------------------------------  
-  
-	COLOR(15,4,4);
-	SCREEN(1);
-
-	VPRINT(0,0,"END");
-	WAIT(30*10);
-   
-}
-
-
-
-/* =============================================================================
-One character input (waiting)
-============================================================================= */
-char INKEY(void) __naked
-{
-__asm 
-  
-	jp BIOS_CHGET
-
-__endasm;
-}
-
-
-
-/* =============================================================================
-  LOCATE
- 
-  Description: 
-            Moves the cursor to the specified location.
-  Input:    (char) Position X of the cursor. (0 to 31 or 79)
-            (char) Position Y of the cursor. (0 to 23)         
-  Output:   -
-============================================================================= */
-void LOCATE(char x, char y) __naked
-{
-x;
-y;
-__asm
-
-  inc  A
-  ld   H,A
-
-  inc  L
-  
-  jp BIOS_POSIT
-
-__endasm;
-}
-
-
-
-//print in screen 1 or 2
-void VPRINT(char column, char line, char* text)
-{
-	uint vaddr;
-	if (GetVDP(1)&0b00010000) vaddr=BASE0 + (line*40)+column;
-	else vaddr=BASE10 + (line*32)+column; // calcula la posicion en la VRAM
-	//VPOKEARRAY(vaddr, text);
-	while(*(text)) VPOKE(vaddr++,*(text++));
-}
-
-
-
-/*void VPOKEARRAY(uint vaddr, char* text)
-{
-  while(*(text)) VPOKE(vaddr++,*(text++));
-}*/
-
-
-
-// Generates a pause in the execution of n interruptions.
-// PAL: 50=1second. ; NTSC: 60=1second. 
-void WAIT(uint cicles)
-{
-  uint i;
-  for(i=0;i<cicles;i++) HALT;
-}
-
-
-
-char PEEK(uint address) __naked
-{
-address;
-__asm
-
-  ld   A,(HL)
-
-  ret
-__endasm;
-}
-
-
-
-void POKE(uint address,char value)
-{
-address;value;
-__asm
-  push IX
-  ld   IX,#0
-  add  IX,SP
-  
-  ld   A,4(IX)
-  ld   (HL),A
-
-  pop  IX
-__endasm;
-}
-
-
-
-void PressKey(void)
-{
-	VPRINT(0,23,text_presskey);
-	LOCATE(25,23);
-	INKEY();	
-}
-
-
-
-void TestVDP(void)
-{
-	char VDPval_before;
-	char VDPval_after;
-	
-	SCREEN(1);
-	VDPval_before = GetVDP(VDP_Mode1);
-	
-	VPRINT(0,0,">Test SetVDP()");
-	WAIT(100);
-	SetVDP(VDP_Mode1,VDPval_before|0b00010000);	//set M1=1
-	HALT;
-	VDPval_after = GetVDP(VDP_Mode1);
-	WAIT(50);
-	SetVDP(VDP_Mode1,VDPval_before);	//restore screen
-	HALT;
-	VPRINT(0,1,">Test GetVDP()");
-	WAIT(50);
-	if (VDPval_after&0b00010000) VPRINT(0,2,">Test=Ok");	//VDPval_before != Textmode1 and VDPval_after == Textmode1
-	else VPRINT(0,2,"Test=ERROR!");
-		
-	PressKey();	
-}
-
-
-
-void TestCLS(void)
-{
-	// ----------------------------------- SCREEN 0	
-	SCREEN(0);
-	
-	FillVRAM(T1_MAP, 0x3C0, 9);	//character 9=circle
-	VPRINT(0,0,">Test CLS() SCREEN 0");
-	TestClear(0);
-	
-	
-	// ----------------------------------- SCREEN 1
-	SCREEN(1);
-	
-	FillVRAM(G1_MAP, 0x300, 203);
-	VPRINT(0,0,">Test CLS() SCREEN 1");
-	TestClear(1);
-}
-
-
-	
-void TestClear(char nscr)
-{	
-	unsigned int BC;
-	unsigned int vaddr;	
-	unsigned int vsize;
-	boolean testResult=true;
-
-	if(nscr==0){
-		vaddr=T1_MAP;
-		vsize=0x3C0;
-	}else{
-		vaddr=G1_MAP;
-		vsize=0x300;
-	}
-			
-
-	WAIT(100);
-	
-	CLS();
-	WAIT(25);
-	
-	for(BC=0;BC<vsize;BC++) if(VPEEK(vaddr++)!=0) testResult=false;
-		
-	WAIT(25);
-	if(testResult==true) VPRINT(0,0,">Test CLS()=Ok");
-	else VPRINT(0,0,">Test CLS()=ERROR!");
-	
-	VPRINT(0,3,text_presskey);
-	LOCATE(25,3);
-	INKEY();
-}
-
-
-
-// TEST SCREEN 0 ###############################################################
-void testSCREEN0(void)
-{
-  char i;
-  
-  COLOR(LIGHT_GREEN,DARK_GREEN,DARK_GREEN);    
-  SCREEN(0);
-
-  VPRINT(31,23," SCREEN 0");
-  WAIT(50);
-  
-  //fillOrdered();
-  for(i=0;i<255;i++) VPOKE(BASE0+i,i);
-  WAIT(50);
-  
-  testVpeekVpoke();
-  WAIT(50);
-  
-  //testFill
-  VPRINT(0,0,"FillVRAM()");
-  for(i=32;i<95;i++)
-  {
-    HALT;
-    HALT;
-    HALT;
-    HALT;
-    
-    FillVRAM (BASE0+10, 942, i);    
-  }
-  
-}
-
-
-
-// TEST SCREEN 1 ###############################################################
-void testSCREEN1(void)
-{
-  COLOR(15,5,1);    
-  SCREEN(1);
-
-  VPRINT(23,23," SCREEN 1");
-  WAIT(50);
-  
-  fillOrdered();
-  WAIT(50);
-  
-  VPRINT(0,10,"Test CopyToVRAM()");  
-  WAIT(50);
-  
-  //copy to VRAM tileset, only gfx patterns
-  CopyToVRAM((uint) TILESET_B0,BASE12+(32*8),96*8);
-  WAIT(50);
-   
-  /* colors
-  The BG Colors array defines 32 colors (each 4 bit background, and 4 bit 
-  foreground, as in VDP register 7). The colors are assigned to the BG Tiles as
-  follows: Tiles 00..07 share the first color, tiles 08..0F share the second 
-  color, etc, and tiles F8..FF share the last color.*/
-  FillVRAM(BASE6,32,0x75);
-  
-  WAIT(50);
-  
-  testFill();
-  
-}
-
-
-
-// TEST SCREEN 2 ###############################################################
-void testSCREEN2(void)
-{
-  COLOR(15,4,1);    
-  SCREEN(2);
-
-  VPRINT(23,23," SCREEN 2");
-  
-  //copy to VRAM tileset, only gfx patterns and fill colors
-  //bank 0
-  HALT;
-  CopyToVRAM((uint) TILESET_B0,BASE12+(32*8),96*8);
-  FillVRAM(BASE11+(32*8),96*8,0x96); //colors (Red)
-    
-  //bank 1
-  CopyToVRAM((uint) TILESET_B0,BASE12+0x800+(32*8),96*8);
-  FillVRAM(BASE11+BANK1+(32*8),96*8,0x3C); //colors (green)
-  
-  //bank 2
-  CopyToVRAM((uint) TILESET_B0,BASE12+0x1000+(32*8),96*8);
-  FillVRAM(BASE11+BANK2+(32*8),96*8,0x54); //colors (blue)
-  WAIT(50);
-  //END copy
-  
-  
-  //test CopyFromVRAM
-  VPRINT(0,5,"Test CopyFromVRAM()");
-  //copy VRAM to RAM
-  CopyFromVRAM(BASE12+(32*8),0xE000,64*8);
-  WAIT(50);
-  
-  //copy RAM to VRAM
-  CopyToVRAM(0xE000,BASE12+0x600,64*8);   
-  FillVRAM(BASE11+0x600,64*8,0xFE); //colors
-  WAIT(120);
-  //END CopyFromVRAM
-  
-  
-  //test fill VRAM  
-  testFill();
-
-}
-
-
-
-// TEST SCREEN 3 ###############################################################
-void testSCREEN3(void)
-{
-  char value=0;
-  uint i;
-  uint vaddr = BASE17;
-  
-  COLOR(15,4,4);
-  SCREEN(1);
-  
-  VPRINT(0,0,"Test SCREEN 3");  
-  WAIT(200);
-  
-     
-  SCREEN(3);
-  
-  for(i=0;i<0x600;i++)
-  {
-    VPOKE(vaddr++,value);
-    if(value<255) value++;
-    else value=0; 
-  }
-  WAIT(250);
-
-}
-
-
-
-
-
-// show tileset
-void fillOrdered(void)
-{
-  char i;
-  for(i=0;i<255;i++) VPOKE(BASE5+i,i);
-}
-
-
-
-// TEST COLOR  #################################################################
-void testColor(void)
-{
-  char i;
-  char inkcol=15;
-
-  COLOR(15,0,4);
-  SCREEN(0);
-
-  VPRINT(15,10,"TEST color");
-  WAIT(50);
-  
-  for(i=0;i<16;i++)
-  { 
-    COLOR(inkcol--,i,i);
-        
-    WAIT(15);
-  }  
-}
-
-
-
-// TEST VPEEK & VPOKE  #########################################################
-// copy a bloq using vpeek and vpoke
-void testVpeekVpoke(void)
-{
-  uint vaddr = BASE0;
-  char i;
-  char value;
-  
-  VPRINT(0,9,"TEST VPEEK & VPOKE:");
-  
-  for(i=0;i<255;i++)
-  {
-    HALT;
-    value = VPEEK(vaddr);
-    VPOKE(vaddr+440,value);
-    vaddr++;  
-  }  
-
-}
-
-
-
-// TEST FILL  ##################################################################
-//fill screen map with a characters range 
-void testFill(void)
-{
-  char i;
-
-  VPRINT(0,0,"FillVRAM()");
-  for(i=32;i<95;i++)
-  {
-    HALT;
-    
-    FillVRAM (BASE10+10, 750, i);
-    
-    HALT;
-    HALT;
-    HALT;
-  }  
-  
-}
-
-
-/* =============================================================================
-   Indicates whether Text 1 mode is active.
-   Output:	1=Yes/True ; 0=No/False
-============================================================================= */
-char isText1Mode(void)
-{
-	//char A = *(unsigned int *) 0xF3E0;	//RG1SAV
-	//if (A&0b00010000) return 1; 		//Text 40col Mode
-	if (GetVDP(1)&0b00010000) return 1; 		//Text 40col Mode
-	return 0;	
-}
-
-
