+++ sources/src/VDP_TMS9918A_MSXBIOS.c
@@ -1,18 +1,19 @@
 /* ============================================================================= 
 # VDP TMS9918A MSX BIOS Library (fR3eL Project)
 
-Version: 1.3 (1/12/2023)
-Author: mvac7/303bcn [mvac7303b@gmail.com] 
-Architecture: MSX
-Format: C object (SDCC .rel)
-Programming language: C and Z80 assembler
-Compiler: SDCC 4.1.12 or newer 
+- Version: 1.4 (12/06/2025)
+- Author: mvac7/303bcn
+- Architecture: MSX
+- Format: C object (SDCC .rel)
+- Programming language: C and Z80 assembler
+- Compiler: SDCC 4.4 or newer 
 
 ## Description:
 Open Source library of functions to work with the TMS9918A video processor
 using BIOS functions 
 
-## History of versions:
+## History of versions (dd/mm/yyyy):
+- v1.4 (12/06/2025) add PUTSPRITE function
 - v1.3 ( 1/12/2023)	update to SDCC (4.1.12) Z80 calling conventions
 - v1.2 (22/12/2020)	Conversion to source in C and added Sprite initialization 
 					functions.
@@ -25,6 +26,8 @@ using BIOS functions
 
 #include "../include/VDP_TMS9918A_MSXBIOS.h"
 
+
+char SPRITE_BF[4];		//buffer with sprite data (Y,X,Pattern number,Color)
 
 
 
@@ -63,98 +66,13 @@ setSCR:
 	ret
 __endasm;
 }    
-
-
-/* =============================================================================
-ClearSprites
-Description: 
-	Initialises the sprite attribute table (OAM). 
-	The vertical location of the sprite is set to 209.
-	The sprite pattern is cleared to null, the sprite number to the sprite 
-	plane number, the sprite colour to the foregtound colour. The vertical 
-	location of the sprite is set to 209 (mode 0 to 3) or 217 (mode 4 to 8).
-Input:	-
-Output:	-
-============================================================================= */
-void ClearSprites(void) __naked
-{
-__asm
-	push IX
-	
-	call BIOS_CLRSPR
-	
-	pop  IX
-	ret
-__endasm;
-} 
-
-
-
-
-/* =============================================================================
-SetSpritesSize
-Description: Set size type for the sprites.
-Input:	[char] size: 0=8x8; 1=16x16
-Output:	-
-============================================================================= */ 
-void SetSpritesSize(char size) __naked
-{
-size;	//A
-__asm
-	ld   HL,#RG1SAV ; --- read vdp(1) from mem
-	ld   B,(HL)
-
-	cp   #1
-	jr   NZ,size8
-	set  1,B ; 16x16
-	jr   setSize
-  
-size8:
-	res  1,B  ; 8x8    
-
-setSize:  
-	ld   C,#0x01
-	jp   BIOS_WRTVDP
-
-__endasm;
-}
-  
-
-
-/* =============================================================================
-SetSpritesZoom
-Description: Set zoom type for the sprites.
-Input:	[char] or [boolean]/[switcher] zoom: 0/false/OFF = x1; 1/true/ON = x2
-Output:	-
-============================================================================= */
-void SetSpritesZoom(char zoom) __naked
-{
-zoom;	//A
-__asm  
-	ld   HL,#RG1SAV ; --- read vdp(1) from mem
-	ld   B,(HL)
-
-	or   A
-	jr   Z,nozoom
-	set  0,B ; zoom
-	jr   setZoom
-  
-nozoom:
-	res  0,B  ;nozoom    
-
-setZoom:  
-	ld   C,#0x01
-	jp   BIOS_WRTVDP
-
-__endasm;
-}
-
   
 
 
 /* =============================================================================
 COLOR
-Description: Specifies the ink, foreground and background colors. 
+Description: 
+		Specifies the ink, foreground, and background colors of the screen 
 Input:	[char] ink color
 		[char] background color
 		[char] border color
@@ -190,8 +108,7 @@ CLS
 Description: 
 		 Clear Screen
 		 Fill the Name Table with the value 0
-		 Does not clear the sprite attribute table (OAM)
-		 
+		 Does not clear the sprite attribute table (OAM)		 
 Input:	-
 Output:	-
 ============================================================================= */
@@ -222,7 +139,8 @@ __endasm;
 
 /* =============================================================================
 VPOKE
-Description: Writes a byte to the video RAM. 
+Description: 
+		Writes a value to VRAM 
 Input:	[unsigned int] VRAM address
 		[char] value
 Output:	- 
@@ -247,7 +165,8 @@ __endasm;
 
 /* =============================================================================
 VPEEK
-Description: Reads data from the video RAM. 
+Description: 
+		Reads a value from VRAM
 Input:	[unsigned int] VRAM address
 Output:	[char] value
 ============================================================================= */ 
@@ -263,7 +182,8 @@ __endasm;
 
 /* =============================================================================
 FillVRAM                               
-Description: Fill a large area of the VRAM of the same byte.
+Description: 
+		Fill a large area of the VRAM of the same byte
 Input:	[unsigned int] address of VRAM
 		[unsigned int] blocklength
 		[char] Value to fill.
@@ -352,8 +272,9 @@ __endasm;
 
 /* =============================================================================
 GetVDP
-Description: 	
-	Provides the mirror value of a VDP register stored in system variables.
+Description:
+		Get value in a VDP register
+		Provides the mirror value of a VDP register stored in system variables
 Input:	[char] VDP register              
 Output:	[char] Value            
 ============================================================================= */
@@ -375,7 +296,7 @@ __endasm;
 /* =============================================================================
 SetVDP
 Description: 
-	Writes a value in VDP registers
+		Writes a value to a VDP register
 Input:	[char] VDP register                     
 		[char] value
 Output:	-             
@@ -388,8 +309,153 @@ __asm
 	ld   C,A ;VDP reg    
 	ld   B,L
 
-	call BIOS_WRTVDP
+	jp   BIOS_WRTVDP
+__endasm;
+}
 
+
+
+
+//############################################################################## SPRITEs functions
+
+
+
+/* =============================================================================
+ClearSprites
+Description: 
+	Initialises the sprite attribute table (OAM). 
+	The vertical location of the sprite is set to 209.
+	The sprite pattern is cleared to null, the sprite number to the sprite 
+	plane number, the sprite colour to the foregtound colour. The vertical 
+	location of the sprite is set to 209 (mode 0 to 3) or 217 (mode 4 to 8).
+Input:	-
+Output:	-
+============================================================================= */
+void ClearSprites(void) __naked
+{
+__asm
+	push IX
+	
+	call BIOS_CLRSPR
+	
+	pop  IX
 	ret
+__endasm;
+} 
+
+
+
+
+/* =============================================================================
+SetSpritesSize
+Description: Set size type for the sprites.
+Input:	[char] size: 0=8x8; 1=16x16
+Output:	-
+============================================================================= */ 
+void SetSpritesSize(char size) __naked
+{
+size;	//A
+__asm
+	ld   HL,#RG1SAV ; --- read vdp(1) from mem
+	ld   B,(HL)
+
+	cp   #1
+	jr   NZ,SetSprSize8
+	set  1,B ; 16x16
+	jr   setVDPREG01
+  
+SetSprSize8:
+	res  1,B  ; 8x8
+	jr   setVDPREG01
+
+__endasm;
+}
+  
+
+
+/* =============================================================================
+SetSpritesZoom
+Description: Set zoom type for the sprites.
+Input:	[char] or [boolean]/[switcher] zoom: 0/false/OFF = x1; 1/true/ON = x2
+Output:	-
+============================================================================= */
+void SetSpritesZoom(char zoom) __naked
+{
+zoom;	//A
+__asm  
+	ld   HL,#RG1SAV ; --- read vdp(1) from mem
+	ld   B,(HL)
+
+	or   A
+	jr   Z,SetSprNOzoom
+	set  0,B ; zoom
+	jr   setVDPREG01
+  
+SetSprNOzoom:
+	res  0,B  ;nozoom    
+
+setVDPREG01:  
+	ld   C,#0x01
+	jp   BIOS_WRTVDP
+
+__endasm;
+}
+
+
+
+/* =============================================================================
+PUTSPRITE
+Description: 
+		Displays a sprite.
+Input:	[char] sprite plane (0-31) 
+		[char] x 
+		[char] y
+		[char] color (0-15)
+		[char] pattern
+Output:	-
+============================================================================= */
+void PUTSPRITE(char plane, char x, char y, char color, char pattern)
+{
+plane;x;y;color;pattern;
+__asm
+	push IX
+	ld   IX,#0
+	add  IX,SP
+
+	ld   E,A	//plane
+	ld   D,L	//x
+	
+	ld   HL,#_SPRITE_BF
+	ld   C,4(IX)
+	ld   (HL),C		//Y
+	inc  HL
+	ld   (HL),D		//X
+	inc  HL
+	ld   C,6(IX)
+//Set sprite pattern number
+//Multiply * 4 when its a 16x16 sprite
+	call BIOS_GSPSIZ          ;0x008A get sprite size in bytes (Carry flag set when size is 16×16)
+	jr   NC,PSsetPat
+
+//IF 16x16
+	SLA  C
+	SLA  C ;x 4 
+	
+PSsetPat:	
+	ld   (HL),C
+	inc  HL
+	ld   C,5(IX)
+	ld   (HL),C		//color
+
+//write Sprite Buffer to VRAM
+	ld   A,E		//plane
+	call BIOS_CALATR	//Returns the address of the sprite attribute table (VRAM) --> HL
+	ld   DE,#_SPRITE_BF
+	ex   DE,HL
+	ld   BC,#4
+	call BIOS_LDIRVM	//Block transfer to VRAM from memory
+	
+	pop  IX
+	
 __endasm;
 }
